Split nanoGPT example into payload, schedule, and driver modules
